Add the remaining collider shapes for debug rendering
